cards profile checks moved to separate validator call
