+++ lib/davinci_crd_test_kit/client/endpoints/mocked_card_responses/launch_smart_app.json
@@ -12,11 +12,6 @@
     }
   },
   "selectionBehavior": "any",
-  "suggestions": [
-    {
-      "label": "Inferno placeholder suggestion"
-    }
-  ],
   "links": [
     {
       "label": "Inferno",
